% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsSubjective.R
\name{downloadOverallSubjectiveDataAsXLSX}
\alias{downloadOverallSubjectiveDataAsXLSX}
\title{Download all subjective data of a study in one xlsx-file}
\usage{
downloadOverallSubjectiveDataAsXLSX(xsServerURL, studyId, apiKey)
}
\arguments{
\item{xsServerURL}{the address of the xs server}

\item{studyId}{the xs-id of the study}

\item{apiKey}{the secret api-key of the study}
}
\description{
Download all subjective data of a study in one xlsx-file
}

