% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsProbands.R
\name{downloadProbands}
\alias{downloadProbands}
\title{Download all proband data of a study}
\usage{
downloadProbands(xsServerURL, studyId, apiKey)
}
\arguments{
\item{xsServerURL}{the address of the xs server}

\item{studyId}{the xs-id of the study}

\item{apiKey}{the secret api-key of the study}
}
\description{
Download all proband data of a study
}

