% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsSubjective.R
\name{downloadSubjectiveDataAsJson}
\alias{downloadSubjectiveDataAsJson}
\title{Download all subjective data of a study's proband}
\usage{
downloadSubjectiveDataAsJson(xsServerURL, studyId, probandId, apiKey)
}
\arguments{
\item{xsServerURL}{the address of the xs server}

\item{studyId}{the xs-id of the study}

\item{probandId}{the xs-id of the proband}

\item{apiKey}{the secret api-key of the study}
}
\description{
Download all subjective data of a study's proband
}

