% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsUnisens.R
\name{downloadUnisensData}
\alias{downloadUnisensData}
\title{Download all Unisens data of a study's proband as a zip file}
\usage{
downloadUnisensData(xsServerURL, studyXSId, probandXSId, apiKey)
}
\arguments{
\item{xsServerURL}{the address of the xs server}

\item{studyXSId}{the xs-id of the study}

\item{probandXSId}{the xs-id of the proband}

\item{apiKey}{the secret api-key of the study}
}
\description{
Download all Unisens data of a study's proband as a zip file
}

