% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsMessaging.R
\name{sendMessageToProband}
\alias{sendMessageToProband}
\title{Send a message to a proband}
\usage{
sendMessageToProband(xsServerURL, studyXSId, probandXSId,
  messageSendingUserEmail, textMessage, apiKey)
}
\arguments{
\item{xsServerURL}{the address of the xs server}

\item{studyXSId}{the study xs-id}

\item{probandXSId}{the proband xs-id}

\item{messageSendingUserEmail}{xs-user that is mentioned as sender}

\item{textMessage}{the message content}

\item{apiKey}{the api-key of the study}
}
\value{
Feedback from the xs server or exception
}
\description{
Send a message to a proband
}

